+++ lib/pleroma/object.ex
@@ -230,4 +230,8 @@ defmodule Pleroma.Object do
       _ -> :noop
     end
   end
+
+  def get_ap_id(%{"id" => id}), do: id
+  def get_ap_id(id) when is_binary(id), do: id
+  def get_ap_id(_), do: {:error, "Object is not a string and has no id."}
 end